requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
bcrypt==3.2.0
bidict==0.23.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.2
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
Mako==1.3.10
MarkupSafe==3.0.2
openai==1.101.0
passlib==1.7.4
pillow==10.4.0
psycopg2==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.0
python-engineio==4.9.1
python-jose==3.3.0
python-multipart==0.0.20
python-socketio==5.11.0
requests==2.32.3
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.40.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.5.0
uvicorn==0.34.2
wsproto==1.2.0
